Refactor: `GasStationPrice` -> `GasStationControl` autosend
The transaction will no be send automatically, and the user will
deal with confirming and/or cancelling it via Metamask's interface
