+++ src/modules/users/components/GasStation/GasStationControls/GasStationControls.css.d.ts
@@ -1,0 +1,2 @@
+export const main: string;
+export const controls: string;
