--- src/modules/users/components/GasStation/GasStationPrice/GasStationPrice.css.d.ts
@@ -1,16 +1,0 @@
-export const main: string;
-export const transactionSpeedContainer: string;
-export const transactionFeeContainer: string;
-export const transactionFeeMenu: string;
-export const transactionSpeedMenuButtonContainer: string;
-export const transactionSpeedMenuButton: string;
-export const transactionFeeActions: string;
-export const transactionFeeInfo: string;
-export const transactionSpeedLabel: string;
-export const transactionFeeLabel: string;
-export const transactionDuration: string;
-export const transactionFeeAmount: string;
-export const transactionFeeConfirm: string;
-export const transactionSpeedContainerToggleable: string;
-export const stateIsSpeedMenuOpen: string;
-export const transactionFeeEthUsd: string;
