--- src/modules/users/components/GasStation/GasStationPrice/index.ts
@@ -1,1 +1,0 @@
-export { default } from './GasStationPrice';
